[ refactor ] 인기 게시글 업데이트 조회 분리

+++ feed/src/main/java/com/linkeleven/msa/feed/infrastructure/client/InteractionClient.java
@@ -19,9 +19,9 @@ public interface InteractionClient {
 
 	// Batch 요청을 위한 API 추가
 	@GetMapping("/external/feeds/comments")
-	Map<Long, Long> getCommentCounts(@RequestParam("feedIds") List<Long> feedIds);
+	Map<Long, Integer> getCommentCounts(@RequestParam("feedIds") List<Long> feedIds);
 
 	@GetMapping("/external/feeds/likes")
-	Map<Long, Long> getLikeCounts(@RequestParam("feedIds") List<Long> feedIds);
+	Map<Long, Integer> getLikeCounts(@RequestParam("feedIds") List<Long> feedIds);
 
 }
